--- src/test-server.js
@@ -1,140 +1,0 @@
-const createError = require('http-errors');
-const express = require('express');
-const cookieParser = require('cookie-parser');
-const logger = require('morgan');
-
-const helmet = require('helmet');
-const cors = require('cors');
-const celebrateErrors = require('celebrate').errors;
-
-
-const app = express();
-
-app.use(logger('dev'));
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-app.use(cookieParser());
-
-app.use(cors());
-app.use(helmet());
-
-app.use('/', router);
-
-app.use(celebrateErrors);
-
-// catch 404 and forward to error handler
-app.use((req, res, next) => {
-  next(createError(404));
-});
-
-// error handler
-app.use((err, req, res) => {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
-
-/**
-* Module dependencies.
-*/
-
-const debug = require('debug')('karma');
-const http = require('http');
-const router = require('./routes/index');
-
-const models = require('./lib/data/models');
-
-/**
-* Get port from environment and store in Express.
-*/
-
-const port = normalizePort(process.env.PORT || '3000');
-app.set('port', port);
-
-/**
-* Create HTTP server.
-*/
-
-const server = http.createServer(app);
-// console.log("Server Created :)");
-
-modelCallback = function () {
-  models.sequelize.sync().then(() => {
-    /**
-        * Listen on provided port, on all network interfaces.
-        */
-    server.on('error', onError);
-    server.on('listening', onListening);
-    server.listen(port);
-  });
-};
-
-models.callback = modelCallback;
-
-
-/**
-* Normalize a port into a number, string, or false.
-*/
-
-function normalizePort(val) {
-  const port = parseInt(val, 10);
-
-  if (isNaN(port)) {
-    // named pipe
-    return val;
-  }
-
-  if (port >= 0) {
-    // port number
-    return port;
-  }
-
-  return false;
-}
-
-/**
-* Event listener for HTTP server "error" event.
-*/
-
-function onError(error) {
-  if (error.syscall !== 'listen') {
-    throw error;
-  }
-
-  const bind = typeof port === 'string'
-    ? `Pipe ${port}`
-    : `Port ${port}`;
-
-    // handle specific listen errors with friendly messages
-  switch (error.code) {
-    case 'EACCES':
-      console.error(`${bind} requires elevated privileges`);
-      process.exit(1);
-      break;
-    case 'EADDRINUSE':
-      console.error(`${bind} is already in use`);
-      process.exit(1);
-      break;
-    default:
-      throw error;
-  }
-}
-
-/**
-* Event listener for HTTP server "listening" event.
-*/
-
-function onListening() {
-  const addr = server.address();
-  const bind = typeof addr === 'string'
-    ? `pipe ${addr}`
-    : `port ${addr.port}`;
-  debug(`Listening on ${bind}`);
-}
-
-
-module.exports = app;
